moving get_signer function to parentchain-signer crate
